+++ cpp/src/io/writer_public.cpp
@@ -65,28 +65,11 @@ Node Writer::DoAddNode(Page &page, VoxelKey key, std::vector<char> in, uint64_t 
     return *node;
 }
 
-Node Writer::AddNode(Page &page, const VoxelKey &key, las::Points &points, bool check_bounds)
+Node Writer::AddNode(Page &page, const VoxelKey &key, las::Points &points)
 {
     auto header = file_->GetLasHeader();
     if (points.PointFormatID() != header.point_format_id || points.PointRecordLength() != header.point_record_length)
         throw std::runtime_error("Writer::AddNode: New points must be of same format and size.");
-
-    if (check_bounds)
-    {
-        // Check that node intersects the las header bounds
-        if (!Box(key, header).Intersects(header.GetBounds()))
-            throw std::runtime_error("Writer::AddNode: Node must intersect the las file bounds");
-        // If node not within las header bounds, check that all points are within the file bounds
-        if (!Box(key, header).Within(header.GetBounds()))
-        {
-            if (!points.Within(Box(header.min, header.max)))
-                throw std::runtime_error("Writer::AddNode: Points must be within the las file bounds");
-        }
-
-        // Check that all points are within the node
-        if (!points.Within(Box(key, header)))
-            throw std::runtime_error("Writer::AddNode: Points must be within the key");
-    }
 
     std::vector<char> uncompressed = points.Pack();
     return AddNode(page, key, uncompressed);