+++ example/example-writer.py
@@ -259,7 +259,7 @@ def NewFileExample():
     key = copc.VoxelKey(0, 0, 0, 0)
     points = RandomPoints(key, header, NUM_POINTS)
     # The node will be written to the file when we call AddNode
-    writer.AddNode(root_page, key, points, check_bounds=True)
+    writer.AddNode(root_page, key, points)
 
     # We can also add pages in the same way, as long as the Key we specify
     # is a child of the parent page
@@ -269,19 +269,23 @@ def NewFileExample():
     # Once our page is created, we can add nodes to it like before
     key = copc.VoxelKey(1, 1, 1, 0)
     points = RandomPoints(key, header, NUM_POINTS)
-    writer.AddNode(page, key, points, check_bounds=True)
+    writer.AddNode(page, key, points)
 
     key = copc.VoxelKey(2, 2, 2, 0)
     points = RandomPoints(key, header, NUM_POINTS)
-    writer.AddNode(page, key, points, check_bounds=True)
+    writer.AddNode(page, key, points)
 
     # We can nest subpages as much as we want, as long as they are children of the parent
     sub_page = writer.AddSubPage(page, copc.VoxelKey(3, 4, 4, 0))
     points = RandomPoints(sub_page.key, header, NUM_POINTS)
-    writer.AddNode(page, sub_page.key, points, check_bounds=True)
+    writer.AddNode(page, sub_page.key, points)
 
     # Make sure we call close to finish writing the file!
     writer.Close()
+
+    # We can check that the spatial bounds of the file have been respected
+    reader = copc.FileReader("new-copc.copc.laz")
+    assert reader.ValidateSpatialBounds()
 
 
 if __name__ == "__main__":