+++ python/bindings.cpp
@@ -420,7 +420,7 @@ PYBIND11_MODULE(copclib, m)
         .def("GetDepthAtResolution", &Reader::GetDepthAtResolution, py::arg("resolution"))
         .def("GetNodesAtResolution", &Reader::GetNodesAtResolution, py::arg("resolution"))
         .def("GetNodesWithinResolution", &Reader::GetNodesWithinResolution, py::arg("resolution"))
-        .def("CheckSpatialBounds", &Reader::CheckSpatialBounds, py::arg("verbose"));
+        .def("ValidateSpatialBounds", &Reader::ValidateSpatialBounds, py::arg("verbose") = false);
 
     py::class_<FileWriter>(m, "FileWriter")
         .def(py::init<const std::string &, Writer::LasConfig const &, const int &, const std::string &>(),
@@ -432,8 +432,8 @@ PYBIND11_MODULE(copclib, m)
         .def("GetExtraByteVlr", &Writer::GetExtraByteVlr)
         .def("GetRootPage", &Writer::GetRootPage)
         .def("Close", &FileWriter::Close)
-        .def("AddNode", py::overload_cast<Page &, const VoxelKey &, las::Points &, bool>(&Writer::AddNode),
-             py::arg("page"), py::arg("key"), py::arg("points"), py::arg("check_bounds") = false)
+        .def("AddNode", py::overload_cast<Page &, const VoxelKey &, las::Points &>(&Writer::AddNode), py::arg("page"),
+             py::arg("key"), py::arg("points"))
         .def("AddNodeCompressed", &Writer::AddNodeCompressed)
         .def("AddNode", py::overload_cast<Page &, const VoxelKey &, std::vector<char> const &>(&Writer::AddNode),
              py::arg("page"), py::arg("key"), py::arg("uncompressed_data"))